Collect import helper needs during module info gathering
