+++ src/compiler/transformers/utilities.ts
@@ -15,6 +15,32 @@ namespace ts {
         hasExportStarsToExportValues: boolean; // whether this module contains export*
     }
 
+    function getNamedImportCount(node: ImportDeclaration) {
+        if (!(node.importClause && node.importClause.namedBindings)) return 0;
+        const names = node.importClause.namedBindings;
+        if (!names) return 0;
+        if (!isNamedImports(names)) return 0;
+        return names.elements.length;
+    }
+
+    function containsDefaultReference(node: NamedImportBindings) {
+        if (!node) return false;
+        if (!isNamedImports(node)) return false;
+        return some(node.elements, isNamedDefaultReference);
+    }
+
+    function isNamedDefaultReference(e: ImportSpecifier) {
+        return e.propertyName && e.propertyName.escapedText === InternalSymbolName.Default;
+    }
+
+    export function getImportNeedsImportStarHelper(node: ImportDeclaration) {
+        return !!getNamespaceDeclarationNode(node) || (getNamedImportCount(node) > 1 && containsDefaultReference(node.importClause.namedBindings));
+    }
+
+    export function getImportNeedsImportDefaultHelper(node: ImportDeclaration) {
+        return isDefaultImport(node) || (getNamedImportCount(node) === 1 && containsDefaultReference(node.importClause.namedBindings));
+    }
+
     export function collectExternalModuleInfo(sourceFile: SourceFile, resolver: EmitResolver, compilerOptions: CompilerOptions): ExternalModuleInfo {
         const externalImports: (ImportDeclaration | ImportEqualsDeclaration | ExportDeclaration)[] = [];
         const exportSpecifiers = createMultiMap<ExportSpecifier>();
@@ -24,6 +50,7 @@ namespace ts {
         let hasExportDefault = false;
         let exportEquals: ExportAssignment = undefined;
         let hasExportStarsToExportValues = false;
+        let hasImportStarOrImportDefault = false;
 
         for (const node of sourceFile.statements) {
             switch (node.kind) {
@@ -33,6 +60,7 @@ namespace ts {
                     // import * as x from "mod"
                     // import { x, y } from "mod"
                     externalImports.push(<ImportDeclaration>node);
+                    hasImportStarOrImportDefault = getImportNeedsImportStarHelper(<ImportDeclaration>node) || getImportNeedsImportDefaultHelper(<ImportDeclaration>node);
                     break;
 
                 case SyntaxKind.ImportEqualsDeclaration:
@@ -135,7 +163,7 @@ namespace ts {
             }
         }
 
-        const externalHelpersModuleName = getOrCreateExternalHelpersModuleNameIfNeeded(sourceFile, compilerOptions, hasExportStarsToExportValues);
+        const externalHelpersModuleName = getOrCreateExternalHelpersModuleNameIfNeeded(sourceFile, compilerOptions, hasExportStarsToExportValues, hasImportStarOrImportDefault);
         const externalHelpersImportDeclaration = externalHelpersModuleName && createImportDeclaration(
             /*decorators*/ undefined,
             /*modifiers*/ undefined,
