+++ tests/baselines/reference/esModuleInteropTslibHelpers.symbols
@@ -1,0 +1,12 @@
+=== tests/cases/compiler/refs.d.ts ===
+declare module "path";
+No type information for this code.=== tests/cases/compiler/file.ts ===
+import path from "path";
+>path : Symbol(path, Decl(file.ts, 0, 6))
+
+path.resolve("", "../");
+>path : Symbol(path, Decl(file.ts, 0, 6))
+
+export class Foo { }
+>Foo : Symbol(Foo, Decl(file.ts, 1, 24))
+
