+++ tests/baselines/reference/esModuleInteropTslibHelpers.types
@@ -1,0 +1,17 @@
+=== tests/cases/compiler/refs.d.ts ===
+declare module "path";
+No type information for this code.=== tests/cases/compiler/file.ts ===
+import path from "path";
+>path : any
+
+path.resolve("", "../");
+>path.resolve("", "../") : any
+>path.resolve : any
+>path : any
+>resolve : any
+>"" : ""
+>"../" : "../"
+
+export class Foo { }
+>Foo : Foo
+
